requirements.txt
customtkinter==5.2.2
pillow==10.3.0
opencv-python-headless==4.9.0.80
matplotlib==3.8.4
numpy==1.24.3